API/requirements.txt
fastapi[standard]
mysql-connector-python
uvicorn[standard]
